+++ ledger/ledger-api-domain/src/main/scala/com/digitalasset/ledger/api/DeduplicationPeriod.scala
@@ -1,9 +1,9 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.v2
+package com.daml.ledger.api
 
-import java.time.Duration
+import java.time.{Duration, Instant}
 
 import com.daml.ledger.offset.Offset
 import com.daml.logging.entries.{LoggingValue, ToLoggingValue}
@@ -17,6 +17,33 @@ import com.daml.logging.entries.{LoggingValue, ToLoggingValue}
 sealed trait DeduplicationPeriod extends Product with Serializable
 
 object DeduplicationPeriod {
+
+  /** Backwards compatibility
+    * Transforms the [[period]] into an [[Instant]] to be used for deduplication into the future.
+    * Offset deduplication is not supported
+    * @param time The time to use for calculating the [[Instant]]. It can either be submission time or current time, based on usage
+    * @param period The deduplication period
+    * @param maxSkew Used when the deduplication period is computed from an [[Instant]].
+    */
+  def deduplicateUntil(
+      time: Instant,
+      period: DeduplicationPeriod,
+      maxSkew: Duration,
+  ): Instant = period match {
+    case DeduplicationDuration(duration) =>
+      time.plus(duration)
+    case DeduplicationOffset(_) =>
+      throw new NotImplementedError("Offset deduplication is not supported")
+    case DeduplicationFromTime(start) =>
+      /** deduplication_start: compute the duration as submissionTime + config.minSkew - deduplication_start
+        * We measure `deduplication_start` on the ledger’s clock, and thus need to add the minSkew to compensate for the maximal skew that the participant might be behind the ledger’s clock.
+        */
+      val duration = Duration.between(
+        time.plus(maxSkew),
+        start,
+      )
+      start.plus(duration)
+  }
 
   /** The length of the deduplication window, which ends when the [[WriteService]] or underlying Daml ledger processes
     * the command submission.
@@ -33,10 +60,14 @@ object DeduplicationPeriod {
   /** The `offset` defines the start of the deduplication period. */
   final case class DeduplicationOffset(offset: Offset) extends DeduplicationPeriod
 
+  final case class DeduplicationFromTime(start: Instant) extends DeduplicationPeriod
+
   implicit val `DeduplicationPeriod to LoggingValue`: ToLoggingValue[DeduplicationPeriod] = {
     case DeduplicationDuration(duration) =>
       LoggingValue.Nested.fromEntries("duration" -> duration)
     case DeduplicationOffset(offset) =>
       LoggingValue.Nested.fromEntries("offset" -> offset)
+    case DeduplicationFromTime(time) =>
+      LoggingValue.Nested.fromEntries("time" -> time)
   }
 }
