+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v2/CompletionInfo.scala
@@ -3,6 +3,7 @@
 
 package com.daml.ledger.participant.state.v2
 
+import com.daml.ledger.api.DeduplicationPeriod
 import com.daml.lf.data.Ref
 import com.daml.logging.entries.{LoggingValue, ToLoggingValue}
