+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v2/Update.scala
@@ -6,6 +6,7 @@ package com.daml.ledger.participant.state.v2
 import java.time.Duration
 
 import com.daml.daml_lf_dev.DamlLf
+import com.daml.ledger.api.DeduplicationPeriod
 import com.daml.ledger.configuration.Configuration
 import com.daml.ledger.grpc.GrpcStatuses
 import com.daml.lf.data.Ref
